requirements.txt
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.7.2
attrs==23.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
coverage==7.4.4
cryptography==41.0.7
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-cors-headers==4.7.0
django-deep-serializer==0.1.3
djangorestframework==3.15.2
entrypoints==0.4
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.46.0
fs==2.4.16
html5lib==1.1
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
ipykernel==6.29.3
ipython==8.20.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
jupyter_client==7.4.9
jupyter_core==5.3.2
keyring==24.3.1
lxml==5.2.1
Markdown==3.5
matplotlib==3.6.3
matplotlib-inline==0.1.6
more-itertools==10.2.0
mysqlclient==2.1.1
nbclient==0.8.0
nbconvert==6.5.3
nbformat==5.9.1
nest-asyncio==1.5.4
notebook==6.4.12
oauthlib==3.2.2
packaging==24.0
paramiko==2.12.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.2.0
platformdirs==3.11.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
Pygments==2.17.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.31.0
rich==13.7.1
SciPy==1.11.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
terminado==0.17.1
tinycss2==1.2.1
tornado==6.4
traitlets==5.5.0
typing_extensions==4.8.0
tzdata==2024.2
urllib3==2.0.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.9
zipp==1.0.0

